requirements.txt
requests
azure-devops
python-dotenv
pytest
ollama